+++ src/main/java/org/bot/nullbot/command/video/VideoDeleteCommand.java
@@ -12,7 +12,6 @@ import org.bot.nullbot.command.Command;
 import org.bot.nullbot.config.prop.FileStorageProperties;
 import org.bot.nullbot.exception.NullBotMsgException;
 import org.bot.nullbot.service.FileService;
-import org.bot.nullbot.util.FileUtil;
 import org.bot.nullbot.util.MessageParseUtil;
 import org.bot.nullbot.util.StringUtil;
 import org.springframework.stereotype.Component;
@@ -31,41 +30,33 @@ public class VideoDeleteCommand implements Command {
 
     @Override
     public void execute(Bot bot, GroupMessageEvent event, List<String> params) {
-        String directory = fileStorageProperties.getVideoPath();
+        String directory = fileStorageProperties.getVideoPath() + "/collect";
         ArrayMsg reply = event.getArrayMsg().getFirst();
+
+        if (!params.isEmpty()) {
+            deleteFile(bot, event, directory, params.getFirst());
+            return;
+        }
         if (reply.getType() == MsgTypeEnum.reply) {
             MsgResp replyMsg = bot.getMsg(reply.getData().get("id").asInt()).getData();
-            // 可优化为单个键值对?
-            Map<String, String> videoMap = MessageParseUtil.parseGroupRawMessageAsVideoMap(replyMsg.getRawMessage());
-            if (videoMap.isEmpty()) throw new NullBotMsgException("[删除视频] ❌未引用视频");
-            for (Map.Entry<String, String> entry : videoMap.entrySet()) {
-                String fileName = entry.getKey();
-                try {
-                    FileUtil.deleteFileByName(directory, fileName);
-                } catch (Exception e) {
-                    throw new NullBotMsgException("[删除视频] ❌" + e.getMessage());
-                }
-                if(!fileService.deleteRecordOnly(directory, fileName))
-                    throw new NullBotMsgException("[删除视频] ❌数据库更新失败");
-                bot.sendGroupMsg(event.getGroupId(), "[删除视频] ⚠️已删除\n- " +
-                        StringUtil.truncateFileName(fileName, 12), false);
-                log.info("\t\t\t\t├─[VideoDelete] 引用视频已删除 - {}", fileName);
-            }
-        } else if (!params.isEmpty()) {
-            String fileName = params.getFirst();
-            try {
-                FileUtil.deleteFileByName(directory, fileName);
-            } catch (Exception e) {
-                throw new NullBotMsgException("[删除视频] ❌" + e.getMessage());
-            }
-            if(!fileService.deleteRecordOnly(directory, fileName))
-                throw new NullBotMsgException("[删除视频] ❌数据库更新失败");
-            bot.sendGroupMsg(event.getGroupId(), "[删除视频] ⚠️已删除\n- " +
-                    StringUtil.truncateFileName(fileName, 12), false);
-            log.info("\t\t\t\t├─[VideoDelete] 指定视频已删除 - {}", fileName);
-        } else {
-            throw new NullBotMsgException("[删除视频] ❌无文件名或引用");
+            Map<String, String> videoMap = MessageParseUtil
+                    .parseGroupRawMessageAsVideoMap(replyMsg.getRawMessage());
+            if (videoMap.isEmpty())
+                throw new NullBotMsgException("[删除视频] ❌未引用视频");
+            for (Map.Entry<String, String> entry : videoMap.entrySet())
+                deleteFile(bot, event, directory, entry.getKey());
+            return;
         }
+
+        throw new NullBotMsgException("[删除视频] ❌无文件名或引用");
+    }
+
+    private void deleteFile(Bot bot, GroupMessageEvent event, String directory, String fileName) {
+        if(!fileService.deleteFile(directory, fileName))
+            throw new NullBotMsgException("[删除视频] ❌失败");
+        bot.sendGroupMsg(event.getGroupId(), "[删除视频] ⚠️已删除\n- " +
+                StringUtil.truncateFileName(fileName, 12), false);
+        log.info("\t\t\t\t├─[VideoDelete] 视频已删除 - {}", fileName);
     }
 
     @Override