+++ CHANGELOG.md
@@ -770,4 +770,7 @@
 - 修复 MessageParseUtil - 将文本 JsonNode 转换为字符串；
 - 优化 WebExceptionHandler - 新增对 Oss 服务相关异常的拦截提示；
 - 优化 CommandListener 和 MonitorListener - 优化结构，方法签名和日志输出；
-- 临时取消 DeepSeekClient 的回显存储 - 似乎会干扰生成格式。
+- 临时取消 DeepSeekClient 的回显存储 - 似乎会干扰生成格式；
+- 弃用 FileUtil 大部分方法 改用 FileService 操作：
+  - FileService 新增数据库和本地合并的文件操作方法并优化相关使用；
+  - FileUtil 构建目录树方法移至 StringUtil。