+++ CHANGELOG.md
@@ -773,4 +773,5 @@
 - 临时取消 DeepSeekClient 的回显存储 - 似乎会干扰生成格式；
 - 弃用 FileUtil 大部分方法 改用 FileService 操作：
   - FileService 新增数据库和本地合并的文件操作方法并优化相关使用；
-  - FileUtil 构建目录树方法移至 StringUtil。
+  - FileUtil 构建目录树方法移至 StringUtil；
+- 其他部分代码优化。